move loader back to preview
